RPART + location
An RPART model fitted on DayOfWeek, PdDistrict and Corner were fitted
with accuracy of only 22%

I also identified that there are 24 times more addresses than crimes, so
addresses will definitely give a clue as to the type of crime that was
committed. However, there are too many levels to make this a good
(efficient, if informative) predictor.